Update forecasting: sklearn DBSCAN for clustering, LSTM for predictions. (#11)
This PR implements a version of Query-Based Workload Forecasting adapted from [QueryBot2000 by Lin et al (2018)](http://www.cs.cmu.edu/~malin199/publications/2018.forecasting.sigmod.pdf). Forecasting is currently divided into three components, each one dependent on artifacts produced by the previous.

**Preprocessor**
This converts a directory of Postgres CSV query logs into a parquet file of templatized queries and the corresponding parameters with which the original query can be reconstructed.

**Clusterer**
Templates are grouped into clusters based on the arrival rate of templates using the [`DBSCAN `](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.DBSCAN.html) algorithm. This produces a mapping from known templates to their assigned clusters.

**Forecaster**
The forecaster uses the templatized queries from the Preprocessor, as well as the assignments from the clustered to produce an aggregate arrival rate time series of each cluster. These rates are used to train an LSTM RNN to predict future arrival rates. The prediction of a cluster's arrival rate, as well as samples from the observed distribution of A.) templates within a query and B.) parameters within a template are used to generate the forecasted workload. This workload is reported in the form of a csv containing (query, count) pairs.

Included in this PR is a sample notebook (`forecast/Sine-Forecast.ipynb`) demonstrating the usage of the clusterer and LSTM model to produce accurate predictions on a synthetic trace of noisy, sinusoidal arrival rates for templates.

In `dodo.py`, forecasting is split into two tasks: `forecast_preprocess` and `forecast_predict`, since it may be the case that preprocess can be done ahead of time before predictions are needed. Currently, `forecast_preprocess` is always run as a dependency if `forecast_predict` is called, but I anticipate there will be cases where we split the forecasting tasks further so the pilot as finer control over the entire process.