ExplainMicroserviceCost Performance Improvements (#37)
ExplainMicroserviceCost Performance Improvements.

This diff addresses issue #35.
Previously with `--record_predictions` on `dev4`, it would take about 1 hour (as recorded by htop)
for the index picker to finish an iteration (and recommend some index). 

This diff introduces the following set of changes:
- Captures the inference time taken per model in seconds. This is only reported as debug information.
- Records the prediction results all at once after processing the entire workload.
- Issues a single model inference request for an entire query explain tree.

With these changes, we get around `<30 minutes` with `--record_predictions` on `dev4` as recorded by htop.