Add Preprocessor support for directly outputting query templates and queries.
This is opt-in behavior gated by `--output-query-templates` and `--output-queries`.

If `--output-queries` is specified, the "query_subst" column of the Preprocessor is kept
even in the final Parquet artifact. If unspecified, existing behavior is maintained.
Storing "query_subst" causes an increase in storage and memory overhead;
expected to be roughly twice as large (previously {time, template, params} were stored).

The goal of this change is to make the Preprocessor useful even without forecasting;
the Preprocessor can now be used to extract all query templates and/or all queries
from query logs with minimal effort, dumping the requested result(s) to CSV.

Also includes other minor code quality fixes (comments, formatting, incomplete docstring).