Plan Differencing Performance Speedup (#52)
Previously it was reported that plan differencing was taking a significant time (for a frame of reference, differencing about 300-400k data points from TPC-C takes about 1 hour on dev4). This refactor significantly speeds up the plan differencing process. A point of reference for the refactored plan differencing:
```
INFO:2022-03-08 00:32:49,088 Mode: train | Benchmark: tpcc_scalefactor_1_terminals_1_rate_10000_time_300
100%|������������������������������| 3633266/3633266 [17:45<00:00, 3409.07it/s]
100%|������������������������������| 9/9 [22:51<00:00, 152.42s/it]
```

The core part of the refactor includes the following changes:
- `<query_id, statement_timestamp, pid>` can uniquely identify a query instance invocation
- Cython module for differencing a given unique query instance invocation
- Removal of `diffed` prefixed target columns from the output
- Removal of most logging constructs

Note:

We currently don't difference parallel query execution. This is done by intentionally not differencing `Gather` and `GatherMerge`. Note that in order to properly difference parallel query execution, we need to (1) ensure that workers record PIDs of the leader and (2) have a reasonable form of "subtracting".